Pin free floating versions
Snapshot versions are replaced as there are new builds. This means that
we don't know exactly what is going into each chart. By pinning the
snapshot version to include a sha, we can lockdown the chart.

This also means that we'll need to update the chart whenever we need to
pick up a change. In time we can automate this process.