Split riff and Istio charts
Because of how Helm manages CRDs, we cannot install Istio and riff at
the same time, they must be separate installs.

There are two Istio charts, istio and istio-lean. These are based on the
values used to generate the Istio installed in Knative. In short, istio
has a full mesh while istio-lean only supports ingress.

The Istio charts are the official Istio charts repackaged with custom
values and versioned with riff.