Maintainable pinned template versions
This adds template file that creates the templates.yaml file. (A bit
meta templating the config that defines how to chart's templates are
defined)

The `{chart}/templates.yaml.tpl` may contain bash expressions which are
resolved and inlined into the output file. The `make templates` target
will update existing chart templates. The output of this target should
be committed to the repo and pass CI before publishing a chart.

Running `make templates` is currently manual, but may be automated in
the future.