Make Istio uniformly consumable
Istio is now optional for both the Core and Knative runtimes. Without
Istio, ingress will not be provided for Knative. For the Core Runtime
ingress can now be provided via Istio.

In theory, this will also make it easier to switch to a different
ingress controller as Knative installs will no longer presume Istio.